Sanitize legacy HTML at the render sink, not just at migration time
Closes #16. The parse5 allow-list added in 8e081ab only ran inside the
migration script, which protected nothing currently served: the rendered
markup comes from src/data/legacy-site-content.json, a committed 6.37 MB
artifact generated before the sanitizer existed, and the script has no
build hook or CI job.

Moves the policy to src/lib/legacy-html.ts and calls it from
LegacyContentPage, immediately before dangerouslySetInnerHTML. The
guarantee now holds regardless of how the markup reached the component,
which matters for two paths the script never touches:

- the committed JSON, which still contains 139 <form> and 3 <iframe>
  elements the old denylist never named
- legacy-content.ts, which synthesizes contentHtml for index pages at
  runtime and never passes through the migration script at all

The migration script now imports the same module instead of carrying its
own copy (-187 lines). Two copies of a security policy drift; the sink and
the generator have to agree by construction, not by discipline.

parse5 moves from devDependencies to dependencies -- it is on the render
path now, not just build tooling.

Verified: next build exits 0 with all routes generated, tsc --noEmit
clean, 27/27 test files pass, sanitizer suite 21/21.

Note this changes rendered output for the legacy pages: forms, iframes and
non-allow-listed wrapper markup are now stripped at render. Measured over
the committed data that is 161 of 165 routes and roughly a third of the
content bytes. Text is preserved -- unknown elements are unwrapped rather
than dropped -- but layout-bearing wrapper markup is not.
