+++ package.json
@@ -21,6 +21,7 @@
     "date-fns": "^3.6.0",
     "lucide-react": "^0.379.0",
     "next": "14.2.35",
+    "parse5": "^8.0.1",
     "react": "^18.3.0",
     "react-dom": "^18.3.0"
   },
@@ -31,7 +32,6 @@
     "autoprefixer": "^10.4.19",
     "eslint": "^8.57.0",
     "eslint-config-next": "14.2.35",
-    "parse5": "^8.0.1",
     "postcss": "^8.5.26",
     "tailwindcss": "^3.4.3",
     "typescript": "^5.4.5"
